move progress worker into mp worker, clean up imports
